SETI-68: Refactor logger and extract it to its own class
